requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
curl_cffi==0.13.0
fastapi==0.134.0
feedparser==6.0.12
fredapi==0.5.2
frozendict==2.4.7
greenlet==3.3.2
h11==0.16.0
idna==3.11
multitasking==0.0.12
numpy==2.4.2
pandas==3.0.1
peewee==4.0.0
platformdirs==4.9.2
praw==7.8.1
prawcore==2.4.0
protobuf==7.34.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.47
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.6.3
uvicorn==0.41.0
websocket-client==1.9.0
websockets==16.0
yfinance==1.2.0
